pyproject.toml
[tool.poetry]
name = "ferrea-datasources"
version = "1.0.0"
description = "Microservice for Ferrea framework."
authors = ["Eugenio Grimoldi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
uvicorn = "^0.27.1"
fastapi-utils = "^0.8.0"
jmespath = "^1.0.1"
ferrea = {git = "https://github.com/Grimoldi/ferrea-lib-chassis.git", rev = "2.2.0"}
dynaconf = "^3.2.11"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-env = "^1.1.5"

[tool.ruff]
# Allow imports relative to the "src" and "test" directories.
src = ["src", "test"]

[tool.mypy]
mypy_path = "src"
--namespace-packages = "True"

[tool.pytest.ini_options]
pythonpath = [
    "./src", 
    "./tests"
]

[tool.pytest_env]
FERREA_APP = "Datasources"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
